+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 *   (improvement) Added `CreativeWork::$position`
 *   (deprecated) Deprecated `Becklyn\SchemaOrg\Schema`, use `Becklyn\SchemaOrg\TypeChecker\TypeChecker` instead.
 *   (feature) Validate nested array values in `TypeChecker` to validate against arrays.
+*   (bug) Update `HowTo::$step`, `::$supply` and `::$tool` to accept an `array` (of various types) or `null`.
 
 
 1.5.0